+++ resources/views/partials/page/head.blade.php
@@ -4,7 +4,7 @@
             <b>{{ $title }}</b>
         </td>
         <td align=right>
-            <a href="{{ route('app_index') }}">Home</a>
+            <a href="{{ route('index') }}">Home</a>
         </td>
     </tr>
 </table>